add folder schemas
