+++ controllers/contacts/index.js
@@ -1,0 +1,13 @@
+const getAll = require("./getAll");
+const getByid = require("./getByid");
+const add = require("./add");
+const updateById = require("./updateById");
+const removeById = require("./removeById");
+
+module.exports = {
+  getAll,
+  getByid,
+  add,
+  updateById,
+  removeById,
+};
