+++ schemas/index.js
@@ -1,0 +1,3 @@
+const addSchema = require("./validation");
+
+module.exports = { addSchema };
